# Conflicts:
#	CHANGELOG-v3.md
#	composer.json
#	src/fields/RichText.php
#	src/web/assets/cp/dist/css/craft.css.map
#	src/web/assets/cp/dist/js/Craft.js
#	src/web/assets/cp/dist/js/Craft.min.js

+++ src/controllers/UsersController.php
@@ -1187,27 +1187,28 @@ class UsersController extends Controller
             $this->requirePermission('editUsers');
         }
 
-        $file = UploadedFile::getInstanceByName('photo');
+        if (($file = UploadedFile::getInstanceByName('photo')) === null) {
+            return null;
+        }
 
         try {
-            // Make sure a file was uploaded
-            if ($file) {
-                if ($file->getHasError()) {
-                    throw new UploadFailedException($file->error);
-                }
-
-                $users = Craft::$app->getUsers();
-                $user = $users->getUserById($userId);
-
-                // Move to our own temp location
-                $fileLocation = Assets::tempFilePath($file->getExtension());
-                move_uploaded_file($file->tempName, $fileLocation);
-                $users->saveUserPhoto($fileLocation, $user, $file->name);
-
-                $html = $this->getView()->renderTemplate('users/_photo', ['account' => $user]);
-
-                return $this->asJson(['html' => $html]);
+            if ($file->getHasError()) {
+                throw new UploadFailedException($file->error);
             }
+
+            $users = Craft::$app->getUsers();
+            $user = $users->getUserById($userId);
+
+            // Move to our own temp location
+            $fileLocation = Assets::tempFilePath($file->getExtension());
+            move_uploaded_file($file->tempName, $fileLocation);
+            $users->saveUserPhoto($fileLocation, $user, $file->name);
+
+            $html = $this->getView()->renderTemplate('users/_photo', ['account' => $user]);
+
+            return $this->asJson([
+                'html' => $html,
+            ]);
         } catch (\Throwable $exception) {
             /** @noinspection UnSafeIsSetOverArrayInspection - FP */
             if (isset($fileLocation)) {
@@ -1216,11 +1217,10 @@ class UsersController extends Controller
 
             Craft::error('There was an error uploading the photo: '.$exception->getMessage(), __METHOD__);
 
-            return $this->asErrorJson(Craft::t('app',
-                'There was an error uploading your photo: {error}', ['error' => $exception->getMessage()]));
+            return $this->asErrorJson(Craft::t('app', 'There was an error uploading your photo: {error}', [
+                'error' => $exception->getMessage()
+            ]));
         }
-
-        return null;
     }
 
     /**
@@ -1711,8 +1711,14 @@ class UsersController extends Controller
         if ($photo = UploadedFile::getInstanceByName('photo')) {
             $fileLocation = Assets::tempFilePath($photo->getExtension());
             move_uploaded_file($photo->tempName, $fileLocation);
-            $users->saveUserPhoto($fileLocation, $user, $photo->name);
-            FileHelper::removeFile($fileLocation);
+            try {
+                $users->saveUserPhoto($fileLocation, $user, $photo->name);
+            } catch (\Throwable $e) {
+                if (file_exists($fileLocation)) {
+                    FileHelper::removeFile($fileLocation);
+                }
+                throw $e;
+            }
         }
     }
